+++ src/register.cpp
@@ -1,5 +1,6 @@
-#include "register.hpp"
 #include <iostream>
+
+#include "register.hpp"
 
 Register::Register(std::string name, Bus* bus): name_(name), bus_(bus), in_(false), out_(false), data_(0) {}
 
@@ -7,19 +8,19 @@ void Register::stepClock()
 {
     if (in_) {
         set_data(bus_->get_data());
-        std::cout << "Write 0x" << std::hex << bus_->get_data() << std::dec << " to " << name_ << std::endl;
+        std::cout << "Write 0x" << std::hex << (int) bus_->get_data() << std::dec << " to " << name_ << std::endl;
     } else if (out_) {
         bus_->set_data(get_data());
-        std::cout << "Read 0x" << std::hex << get_data() << std::dec << " from " << name_ << std::endl;
+        std::cout << "Read 0x" << std::hex << (int) get_data() << std::dec << " from " << name_ << std::endl;
     }
 }
 
-int Register::get_data() const
+uint8_t Register::get_data() const
 {
     return data_;
 }
 
-void Register::set_data(int data)
+void Register::set_data(uint8_t data)
 {
     data_ = data;
 }
